Change libflake API to take 32-bit integer samples instead of 16-bit integer samples. To do this cleanly, modify the PCM reader to output based on selected data format instead of sample format. Also, remove floating-point support from PCM reader.

+++ libflake/encode.h
@@ -93,6 +93,6 @@ typedef struct FlacEncodeContext {
 } FlacEncodeContext;
 
 extern int encode_frame(FlacEncodeContext *s, uint8_t *frame_buffer,
-                        int buf_size, const int16_t *samples, int block_size);
+                        int buf_size, const int32_t *samples, int block_size);
 
 #endif /* FLAC_H */
